+++ src/main/java/ch/njol/skript/events/EvtVehicleCollision.java
@@ -7,8 +7,11 @@ import ch.njol.skript.lang.Literal;
 import ch.njol.skript.lang.SkriptEvent;
 import ch.njol.skript.lang.SkriptParser.ParseResult;
 import ch.njol.skript.lang.SyntaxStringBuilder;
+import ch.njol.skript.registrations.EventValues;
+import ch.njol.skript.util.Getter;
 import org.bukkit.block.Block;
 import org.bukkit.block.data.BlockData;
+import org.bukkit.entity.Entity;
 import org.bukkit.event.Event;
 import org.bukkit.event.vehicle.VehicleBlockCollisionEvent;
 import org.bukkit.event.vehicle.VehicleCollisionEvent;
@@ -23,11 +26,27 @@ public class EvtVehicleCollision extends SkriptEvent {
 	static {
 		Skript.registerEvent("Vehicle Collision", EvtVehicleCollision.class, new Class[]{VehicleBlockCollisionEvent.class, VehicleEntityCollisionEvent.class},
 				"vehicle collision [(with|of) [a[n]] %-itemtypes/blockdatas/entitydatas%]",
-			"vehicle block collision [(with|of) [a[n]] %-itemtypes/blockdatas%]",
-			"vehicle entity collision [(with|of) [a[n]] %-entitydatas%]")
+				"vehicle block collision [(with|of) [a[n]] %-itemtypes/blockdatas%]",
+				"vehicle entity collision [(with|of) [a[n]] %-entitydatas%]")
 				.description("Called when a vehicle collides with a block or entity.")
 				.examples("on vehicle collision:", "on vehicle collision with obsidian:", "on vehicle collision with a zombie:")
 				.since("INSERT VERSION");
+
+		// VehicleBlockCollisionEvent
+		EventValues.registerEventValue(VehicleBlockCollisionEvent.class, Block.class, new Getter<>() {
+			@Override
+			public Block get(VehicleBlockCollisionEvent event) {
+				return event.getBlock();
+			}
+		}, EventValues.TIME_NOW);
+
+		// VehicleEntityCollisionEvent
+		EventValues.registerEventValue(VehicleEntityCollisionEvent.class, Entity.class, new Getter<>() {
+			@Override
+			public Entity get(VehicleEntityCollisionEvent event) {
+				return event.getEntity();
+			}
+		}, EventValues.TIME_NOW);
 	}
 
 	private Literal<?> expr;
@@ -41,7 +60,7 @@ public class EvtVehicleCollision extends SkriptEvent {
 	public boolean init(Literal<?>[] args, int matchedPattern, ParseResult parseResult) {
 		if (args[0] != null) {
 			expr = args[0];
-			for (Object object : expr.getArray()) {
+			for (Object object : expr.getAll()) {
 				if (object instanceof ItemType itemType) {
 					itemTypes.add(itemType);
 				} else if (object instanceof BlockData blockData) {
@@ -96,10 +115,11 @@ public class EvtVehicleCollision extends SkriptEvent {
 	public String toString(@Nullable Event event, boolean debug) {
 		SyntaxStringBuilder builder = new SyntaxStringBuilder(event, debug);
 		builder.append("vehicle");
-		if (blockCollision)
+		if (blockCollision) {
 			builder.append("block");
-		else if (entityCollision)
+		} else if (entityCollision) {
 			builder.append("entity");
+		}
 		builder.append("collision");
 		if (expr != null)
 			builder.append("of", expr);
